include as part of `LocalServerDescriptor` and `NetworkDescriptor`
