+++ src/dfx-core/src/network/mod.rs
@@ -1,4 +1,3 @@
 pub mod directory;
 pub mod provider;
 pub mod root_key;
-pub mod uri;
